+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * `RedisService`'s new `set_state` method to easily set the state of the internal redis DB.
 * automatic linting on every PR
 * all extra client methods also accept `**kwargs` forwarded to client constructor
+* `BlobStorageService.connection_string` for obtaining the connection string to the container's blob storage.
 ### Fixed
 * Bug where tests failed on linux because linux doesn't have `host.docker.internal`. Linux now uses IP `172.17.0.1` to target host.
 * Bug where services would fail if trying to run a non-pulled image.